hitech_bank/requirements.txt
flask
flask-cors
python-dotenv
pynacl
firebase-admin
requests
psycopg2-binary
PyJWT
